// libs/report-states/ui/src/lib/adapters/approve-report-states-vm-props.interface.ts
import { TableRowBase } from '@cmz/shared-ui';

export interface ApproveReportStatesVmProps extends TableRowBase {
    uniqId: string;
    reportTypeLabel: string;
    operatorsLabel: string;
    sourceLabel: string;
    initiatorPhoneNumber: string;
    reportedAt: string;
    actionsRef: string;
}